FXVPN-381: Perform DNS resolution in-process with c-ares (#10304) (#10316)
Many domains resolve their hostnames into addresses which depend on the user's location,
meaning that to truly exclude an application we also need to exclude their DNS queries. In
order to do this correctly, we must make use a nameserver outside the VPN tunnel.

Unfortunately, this is difficult, if not impossible to accomplish with a `QDnsLookup` as it often
uses in-kernel resolvers and has difficulty with the Windows killswitch. Instead we make use
of the c-ares library, which ensures that it can take place in the same process as the socks
proxy.

---------

Co-authored-by: Naomi Kirby <oskirby@gmail.com>
Co-authored-by: Sebastian Streich <streich.mobile@gmail.com>