Add setuptools to requirements.in
The glean parser script uses `pkg_resources`, and according to the release notes for Python 3.12[^1], `setuptools` is not included in a virtual environment by default any more.  So include it here.  I don't know what would be the appropriate version, the latest version worked for me.  So I used that.

[^1]: https://docs.python.org/3/whatsnew/3.12.html