Prevent multiple emissions of Controller::readyToQuit
There is a potential bug in the Controller class that can cause readyToQuit
to be sent multiple times. For example, this can occur when the controller
is disconnected, we set m_nextStep to Quit and then an asynchronous
disconnected signal is sent by the implementation (eg: on background service
error).

To prevent this from occurring, we should only set m_nextStep when we are
expecting a disconnected signal