Make Daemon::cleanLogs() Q_INVOKABLE and fix crash in XPC log cleanup
When cleaning up the daemon logs via XPC, we have a use-after-free bug
with the transient XPC connection. What occurs is that the lambda capture
holds a reference to self, which becomes invalid as the transient XPC
connection is closed, and then gets used later in the Daemon's Qt thread.