VPN-7328 - Report Android Manifacturer for Support
We get reports that on some android devices disconnecting even with 
always-on. Logs don't show any specific bug or any dissconnection caused by us 
or the network. 

So we might be fighting with background closing issues. 

In order to better understand that, support should know the device that is failing
so we can point to the manufacturer.